refactor: Always return results from `inplace` operations. (#232)
Updates operations that could be performed inplace to always return an ScmRun. The inplace argument dictates whether it is a copy or not.

This also fixes a typing issue for run_append with a return type of BaseScmRun | None. It now returns the type of the items that are being appended.

This also drops pandas 1.0.5 support. 

Closes #230